refactor(dashboard): move review-inbox cards to the deck grid column
Split the hanging sign's content from its "study all" action: the sign
now hosts a placeholder actions panel (deck actions land here later)
while the review-inbox card carousel moves above the deck grid in the
right column, unchanged in styling/behavior.

Rename inbox-toggle animation utils to dashboard-actions since the sign
no longer swings in review-inbox content.
