+++ src/utils/animations/dashboard-actions.ts
@@ -1,0 +1,49 @@
+import { gsap } from 'gsap'
+
+const ENTER_DURATION = 0.35
+const LEAVE_DURATION = 0.2
+
+export function actionsSwingBeforeEnter(el: Element) {
+  const wrapper = el as HTMLElement
+  const panel = wrapper.firstElementChild as HTMLElement
+  gsap.set(wrapper, { height: 0, overflow: 'hidden' })
+  gsap.set(panel, { rotateX: -90, opacity: 0, transformOrigin: 'top center' })
+}
+
+export function actionsSwingEnter(el: Element, done: () => void) {
+  const wrapper = el as HTMLElement
+  const panel = wrapper.firstElementChild as HTMLElement
+  const h = panel.offsetHeight
+
+  gsap.to(wrapper, { height: h, duration: ENTER_DURATION * 0.75, ease: 'power3.out' })
+  gsap.to(panel, {
+    rotateX: 0,
+    opacity: 1,
+    duration: ENTER_DURATION,
+    ease: 'back.out(1.6)',
+    onComplete: () => {
+      gsap.set(wrapper, { height: 'auto', overflow: '' })
+      done()
+    }
+  })
+}
+
+export function actionsSwingLeave(el: Element, done: () => void) {
+  const wrapper = el as HTMLElement
+  const panel = wrapper.firstElementChild as HTMLElement
+
+  gsap.set(wrapper, { height: wrapper.offsetHeight, overflow: 'hidden' })
+  gsap.to(panel, {
+    rotateX: -90,
+    opacity: 0,
+    transformOrigin: 'top center',
+    duration: LEAVE_DURATION * 0.65,
+    ease: 'power2.in'
+  })
+  gsap.to(wrapper, {
+    height: 0,
+    duration: LEAVE_DURATION,
+    ease: 'power3.in',
+    onComplete: done
+  })
+}
